+++ config/prod.exs
@@ -13,11 +13,5 @@ config :swoosh, api_client: Swoosh.ApiClient.Finch, finch_name: Deployex.Finch
 # Disable Swoosh Local Memory Storage
 config :swoosh, local: false
 
-# Runtime production configuration, including reading
-# of environment variables, is done on config/runtime.exs.
-
-config :deployex,
-  base_path: "/var/lib/deployex"
-
 # Do not print debug messages in production
 config :logger, level: :info